requirements.txt
# ═══ V7LTHRONYX VPN Panel v2.0 - Python Dependencies ═══

# Core Framework (FastAPI)
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Database (PostgreSQL)
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg[binary]>=3.1.0
asyncpg>=0.29.0

# Caching & Rate Limiting
redis>=5.0.0

# Security & Crypto
argon2-cffi>=21.3.0
pyotp>=2.8.0
python-jose[cryptography]>=3.3.0
cryptography>=41.0.0
pycryptodome>=3.18.0
passlib>=1.7.4

# Background Jobs
celery>=5.3.0

# System & Process
psutil>=5.9.0

# HTTP Client (Xray API calls)
requests>=2.31.0
httpx>=0.25.0

# ═══ Protocol Dependencies ═══

# WireGuard / AmneziaWG
python-wireguard>=0.1.0

# OpenVPN management
openvpn-api>=0.1.0

# DPI Evasion (scapy + aiohttp)
scapy>=2.5.0
aiohttp>=3.9.0

# NOTE: Hysteria2, TUIC, NaiveProxy, ShadowTLS, and Mieru are managed
# via their own CLI binaries (not Python packages). The panel generates
# config files and controls them through subprocess calls to the
# installed binaries — no pip SDK needed.
